refactor adventure module
