+++ src/game/utils.js
@@ -1,0 +1,22 @@
+import { clamp } from './engine.js';
+
+export function setText(id, v) {
+  const el = document.getElementById(id);
+  if (el) el.textContent = v;
+}
+
+export function setFill(id, ratio) {
+  ratio = clamp(ratio, 0, 1);
+  const el = document.getElementById(id);
+  if (el) el.style.width = (ratio * 100).toFixed(1) + '%';
+}
+
+export function log(msg, cls = '') {
+  const el = document.getElementById('log');
+  if (!el) return;
+  const p = document.createElement('p');
+  if (cls) p.className = cls;
+  p.textContent = msg;
+  el.appendChild(p);
+  el.scrollTop = el.scrollHeight;
+}
