+++ ui/index.js
@@ -21,6 +21,17 @@ import {
   calculatePlayerAttackRate
 } from '../src/game/engine.js';
 import { initHp } from '../src/game/helpers.js';
+import { setText, setFill, log } from '../src/game/utils.js';
+import {
+  ADVENTURE_ZONES,
+  updateActivityAdventure,
+  updateAdventureCombat,
+  startAdventureCombat,
+  progressToNextArea,
+  retreatFromCombat,
+  updateFistProficiencyDisplay,
+  updateFoodSlots
+} from '../src/game/adventure.js';
 
 // Global variables
 let selectedActivity = 'cultivation'; // Current selected activity for the sidebar
@@ -356,86 +367,6 @@ import { ENEMY_DATA } from '../data/enemies.js';
 
 // Adventure System Data
 // Enemy data for adventure zones
-const ADVENTURE_ZONES = [
-  {
-    name: 'Peaceful Lands',
-    description: 'A serene area perfect for beginners',
-    unlockReq: { zone: 0, area: 0 },
-    areas: [
-      { name: 'Forest Edge', enemy: 'Forest Rabbit', killReq: 5 },
-      { name: 'Meadow Path', enemy: 'Wild Boar', killReq: 8 },
-      { name: 'Creek Crossing', enemy: 'River Frog', killReq: 10 },
-      { name: 'Flower Field', enemy: 'Honey Bee', killReq: 12 },
-      { name: 'Old Oak Grove', enemy: 'Tree Sprite', killReq: 15 },
-      { name: 'Mossy Rocks', enemy: 'Stone Lizard', killReq: 18 },
-      { name: 'Babbling Brook', enemy: 'Water Snake', killReq: 20 },
-      { name: 'Sunny Clearing', enemy: 'Grass Wolf', killReq: 25 },
-      { name: 'Ancient Ruins', enemy: 'Ruin Guardian', killReq: 30 },
-      { name: 'Forest Heart', enemy: 'Forest Spirit', killReq: 40 }
-    ]
-  },
-  {
-    name: 'Dark Woods',
-    description: 'Twisted trees hide dangerous creatures',
-    unlockReq: { zone: 0, area: 10 },
-    areas: [
-      { name: 'Shadow Path', enemy: 'Shadow Wolf', killReq: 50 },
-      { name: 'Twisted Grove', enemy: 'Dark Treant', killReq: 60 },
-      { name: 'Cursed Pond', enemy: 'Cursed Toad', killReq: 70 },
-      { name: 'Thorn Thicket', enemy: 'Thorn Beast', killReq: 80 },
-      { name: 'Witch\'s Hut', enemy: 'Corrupted Familiar', killReq: 90 },
-      { name: 'Dead Tree Circle', enemy: 'Wraith', killReq: 100 },
-      { name: 'Nightmare Clearing', enemy: 'Nightmare Hound', killReq: 120 },
-      { name: 'Bone Yard', enemy: 'Skeleton Warrior', killReq: 140 },
-      { name: 'Dark Altar', enemy: 'Dark Cultist', killReq: 160 },
-      { name: 'Shadow Lord\'s Lair', enemy: 'Shadow Lord', killReq: 200 }
-    ]
-  },
-  {
-    name: 'Mountain Peaks',
-    description: 'High altitude challenges await the brave',
-    unlockReq: { zone: 1, area: 10 },
-    areas: [
-      { name: 'Rocky Trail', enemy: 'Mountain Goat', killReq: 250 },
-      { name: 'Wind Caves', enemy: 'Wind Elemental', killReq: 300 },
-      { name: 'Ice Fields', enemy: 'Frost Bear', killReq: 350 },
-      { name: 'Crystal Cavern', enemy: 'Crystal Golem', killReq: 400 },
-      { name: 'Eagle\'s Nest', enemy: 'Giant Eagle', killReq: 450 },
-      { name: 'Storm Peak', enemy: 'Lightning Hawk', killReq: 500 },
-      { name: 'Avalanche Zone', enemy: 'Ice Titan', killReq: 600 },
-      { name: 'Dragon\'s Perch', enemy: 'Young Dragon', killReq: 700 },
-      { name: 'Sky Temple', enemy: 'Sky Guardian', killReq: 800 },
-      { name: 'Peak Summit', enemy: 'Mountain King', killReq: 1000 }
-    ]
-  }
-];
-
-const ADVENTURE_ENEMIES = {
-  // Zone 1: Peaceful Lands
-  'Forest Rabbit': { hp: 25, attack: 3, attackRate: 0.8, loot: { spiritStones: 2, wood: 1 }, meatDrop: { chance: 0.6, amount: 1 } },
-  'Wild Boar': { hp: 45, attack: 8, attackRate: 0.6, loot: { spiritStones: 4, wood: 2 }, meatDrop: { chance: 0.8, amount: 2 } },
-  'Forest Wolf': { hp: 60, attack: 12, attackRate: 0.9, loot: { spiritStones: 6, ironOre: 1 }, meatDrop: { chance: 0.7, amount: 1 } },
-  'Bear': { hp: 100, attack: 18, attackRate: 0.5, loot: { spiritStones: 10, wood: 3 }, meatDrop: { chance: 0.9, amount: 3 } },
-  'Forest Guardian': { hp: 150, attack: 25, attackRate: 0.7, loot: { spiritStones: 15, wood: 5, ironOre: 2 } }, // No meat (magical creature)
-  
-  'Bandit': { hp: 35, attack: 6, attackRate: 0.7, loot: { spiritStones: 3, ironOre: 1 } }, // No meat (human)
-  'Bandit Leader': { hp: 80, attack: 15, attackRate: 0.8, loot: { spiritStones: 8, ironOre: 2 } }, // No meat (human)
-  'Rogue Cultivator': { hp: 120, attack: 22, attackRate: 1.0, loot: { spiritStones: 12, ironOre: 3 } }, // No meat (human)
-  'Mercenary': { hp: 90, attack: 20, attackRate: 0.9, loot: { spiritStones: 10, ironOre: 2 } }, // No meat (human)
-  'Bandit King': { hp: 200, attack: 35, attackRate: 1.1, loot: { spiritStones: 25, ironOre: 5 } }, // No meat (human)
-  
-  'Stone Golem': { hp: 180, attack: 30, attackRate: 0.4, loot: { spiritStones: 18, ironOre: 4 } }, // No meat (construct)
-  'Crystal Spider': { hp: 90, attack: 25, attackRate: 1.2, loot: { spiritStones: 12, iceCrystal: 2 } }, // No meat (insect)
-  'Cave Troll': { hp: 250, attack: 40, attackRate: 0.6, loot: { spiritStones: 30, ironOre: 6 }, meatDrop: { chance: 0.5, amount: 2 } },
-  'Earth Elemental': { hp: 300, attack: 45, attackRate: 0.5, loot: { spiritStones: 35, ironOre: 8, iceCrystal: 3 } }, // No meat (elemental)
-  'Mountain Dragon': { hp: 500, attack: 70, attackRate: 0.8, loot: { spiritStones: 60, ironOre: 12, iceCrystal: 6 }, meatDrop: { chance: 1.0, amount: 5 } },
-  
-  'Skeleton Warrior': { hp: 120, attack: 28, attackRate: 0.7, loot: { spiritStones: 15, ironOre: 3 } }, // No meat (undead)
-  'Lich': { hp: 200, attack: 50, attackRate: 1.0, loot: { spiritStones: 25, iceCrystal: 4 } }, // No meat (undead)
-  'Ancient Guardian': { hp: 350, attack: 60, attackRate: 0.6, loot: { spiritStones: 40, ironOre: 8, iceCrystal: 5 } }, // No meat (construct)
-  'Cursed Spirit': { hp: 180, attack: 45, attackRate: 1.1, loot: { spiritStones: 22, iceCrystal: 3 } }, // No meat (spirit)
-  'Ruin Lord': { hp: 600, attack: 85, attackRate: 0.9, loot: { spiritStones: 75, ironOre: 15, iceCrystal: 10 } }, // No meat (undead lord)
-};
 
 function updateQiOrbEffect(){
   const qiOrb = document.getElementById('qiOrb');
@@ -644,8 +575,6 @@ function updateAll(){
   updateActivityCards();
 }
 
-function setText(id, v){const el=document.getElementById(id); if(el) el.textContent=v}
-function setFill(id, ratio){ratio=clamp(ratio,0,1); const el=document.getElementById(id); if(el) el.style.width=(ratio*100).toFixed(1)+'%'}
 
 function renderUpgrades(){
   // Legacy function - now handled by renderBuildings in sect tab
@@ -960,453 +889,6 @@ function selectActivity(activityType) {
 
 // Combat calculation functions
 
-function gainFistXP(amount) {
-  if (!S.proficiencies) return;
-  const prof = S.proficiencies.fist;
-  prof.exp += amount;
-  while (prof.exp >= prof.expMax) {
-    prof.exp -= prof.expMax;
-    prof.level++;
-    prof.expMax = Math.floor(prof.expMax * 1.5);
-    log(`Fist proficiency reached level ${prof.level}!`, 'good');
-  }
-  updateFistProficiencyDisplay();
-}
-
-function updateFistProficiencyDisplay() {
-  if (!S.proficiencies) return;
-  const prof = S.proficiencies.fist;
-  setText('fistLevel', prof.level);
-  setText('fistExp', Math.floor(prof.exp));
-  setText('fistExpMax', prof.expMax);
-  setFill('fistExpFill', prof.exp / prof.expMax);
-}
-
-// Adventure zone and area UI functions
-function updateZoneButtons() {
-  // Initialize adventure data if needed
-  if (!S.adventure) {
-    S.adventure = {
-      currentZone: 0,
-      currentArea: 0,
-      selectedZone: 0,
-      selectedArea: 0,
-      totalKills: 0,
-      areasCompleted: 0,
-      zonesUnlocked: 1,
-      killsInCurrentArea: 0,
-      inCombat: false
-    };
-  }
-  
-  // Update zone selection buttons (if they exist in the UI)
-  const zoneContainer = document.getElementById('zoneButtons');
-  if (zoneContainer && ADVENTURE_ZONES) {
-    zoneContainer.innerHTML = '';
-    ADVENTURE_ZONES.forEach((zone, index) => {
-      if (index < (S.adventure.zonesUnlocked || 1)) {
-        const button = document.createElement('button');
-        button.className = 'btn zone-btn' + (index === (S.adventure.selectedZone || 0) ? ' active' : '');
-        button.textContent = zone.name;
-        button.onclick = () => selectZone(index);
-        zoneContainer.appendChild(button);
-      }
-    });
-  }
-}
-
-function updateAreaGrid() {
-  // Initialize adventure data if needed
-  if (!S.adventure) return;
-  
-  // Update area selection grid (if it exists in the UI)
-  const areaContainer = document.getElementById('areaGrid');
-  if (areaContainer && ADVENTURE_ZONES) {
-    const currentZone = ADVENTURE_ZONES[S.adventure.selectedZone || 0];
-    if (currentZone && currentZone.areas) {
-      areaContainer.innerHTML = '';
-      currentZone.areas.forEach((area, index) => {
-        const button = document.createElement('button');
-        button.className = 'btn area-btn' + (index === (S.adventure.selectedArea || 0) ? ' active' : '');
-        button.textContent = area.name;
-        button.onclick = () => selectArea(index);
-        areaContainer.appendChild(button);
-      });
-    }
-  }
-}
-
-
-
-
-
-// Battle display function
-function updateBattleDisplay() {
-  // Initialize adventure data if needed
-  if (!S.adventure) {
-    const { hp: enemyHP, hpMax: enemyMaxHP } = initHp(0);
-    S.adventure = {
-      currentZone: 0,
-      currentArea: 0,
-      selectedZone: 0,
-      selectedArea: 0,
-      totalKills: 0,
-      areasCompleted: 0,
-      zonesUnlocked: 1,
-      killsInCurrentArea: 0,
-      inCombat: false,
-      playerHP: S.hp || 100,
-      enemyHP,
-      enemyMaxHP,
-      currentEnemy: null,
-      combatLog: []
-    };
-  }
-  
-  // Update player HP display
-  const playerHP = S.adventure.playerHP || S.hp || 100;
-  const playerMaxHP = S.hpMax || 100;
-  setText('playerHealthText', `${playerHP}/${playerMaxHP}`);
-  
-  // Update player health bar
-  const playerHealthFill = document.getElementById('playerHealthFill');
-  if (playerHealthFill) {
-    const playerHealthPct = (playerHP / playerMaxHP) * 100;
-    playerHealthFill.style.width = `${playerHealthPct}%`;
-  }
-  
-  // Update player attack stats
-  const playerAttack = calculatePlayerCombatAttack();
-  const playerAttackRate = calculatePlayerAttackRate();
-  setText('playerAttack', Math.floor(playerAttack));
-  setText('playerAttackRate', `${playerAttackRate.toFixed(1)}/s`);
-  
-  // Update enemy display if in combat
-  if (S.adventure.inCombat && S.adventure.currentEnemy) {
-    const enemy = S.adventure.currentEnemy;
-    const enemyHP = S.adventure.enemyHP || 0;
-    const enemyMaxHP = S.adventure.enemyMaxHP || 0;
-    
-    setText('enemyName', enemy.name || 'Unknown Enemy');
-    setText('enemyHealthText', `${enemyHP}/${enemyMaxHP}`);
-    setText('enemyAttack', enemy.attack || 0);
-    setText('enemyAttackRate', `${(enemy.attackRate || 1.0).toFixed(1)}/s`);
-    
-    // Update enemy health bar
-    const enemyHealthFill = document.getElementById('enemyHealthFill');
-    if (enemyHealthFill && enemyMaxHP > 0) {
-      const enemyHealthPct = (enemyHP / enemyMaxHP) * 100;
-      enemyHealthFill.style.width = `${enemyHealthPct}%`;
-    }
-  } else {
-    // Show default enemy display when not in combat
-    setText('enemyName', 'Select an area to begin');
-    setText('enemyHealthText', '--/--');
-    setText('enemyAttack', '--');
-    setText('enemyAttackRate', '--/s');
-    
-    // Reset enemy health bar
-    const enemyHealthFill = document.getElementById('enemyHealthFill');
-    if (enemyHealthFill) {
-      enemyHealthFill.style.width = '0%';
-    }
-  }
-  
-  // Update combat log if it exists
-  const combatLog = document.getElementById('combatLog');
-  if (combatLog && S.adventure.combatLog) {
-    // Show last 5 combat log entries
-    const recentLogs = S.adventure.combatLog.slice(-5);
-    combatLog.innerHTML = recentLogs.map(log => `<div class="log-entry">${log}</div>`).join('');
-    combatLog.scrollTop = combatLog.scrollHeight;
-  }
-}
-
-// Adventure combat system functions
-function updateAdventureCombat() {
-  // Initialize adventure data if needed
-  if (!S.adventure) return;
-  
-  // Only update combat if we're in combat
-  if (!S.adventure.inCombat) return;
-  
-  // Basic combat logic - this is a placeholder for the full combat system
-  if (S.adventure.currentEnemy && S.adventure.enemyHP > 0) {
-    // Combat is ongoing
-    const playerAttack = calculatePlayerCombatAttack();
-    const playerAttackRate = calculatePlayerAttackRate();
-    
-    // Simple combat tick (this would be more complex in a full implementation)
-    const now = Date.now();
-    if (!S.adventure.lastPlayerAttack) S.adventure.lastPlayerAttack = now;
-    if (!S.adventure.lastEnemyAttack) S.adventure.lastEnemyAttack = now;
-    
-    // Player attacks
-    if (now - S.adventure.lastPlayerAttack >= (1000 / playerAttackRate)) {
-      S.adventure.enemyHP = Math.max(0, S.adventure.enemyHP - playerAttack);
-      S.adventure.lastPlayerAttack = now;
-
-      // Gain fist proficiency experience for each attack
-      gainFistXP(playerAttack);
-
-      if (!S.adventure.combatLog) S.adventure.combatLog = [];
-      S.adventure.combatLog.push(`You deal ${playerAttack} damage to ${S.adventure.currentEnemy.name}`);
-      
-      // Check if enemy is defeated
-      if (S.adventure.enemyHP <= 0) {
-        defeatEnemy();
-      }
-    }
-    
-    // Enemy attacks back (if still alive)
-    if (S.adventure.enemyHP > 0 && S.adventure.currentEnemy) {
-      const enemyAttackRate = S.adventure.currentEnemy.attackRate || 1.0;
-      if (now - S.adventure.lastEnemyAttack >= (1000 / enemyAttackRate)) {
-        const enemyDamage = S.adventure.currentEnemy.attack || 5;
-        S.adventure.playerHP = Math.max(0, S.adventure.playerHP - enemyDamage);
-        S.adventure.lastEnemyAttack = now;
-        
-        S.adventure.combatLog.push(`${S.adventure.currentEnemy.name} deals ${enemyDamage} damage to you`);
-        
-        // Check if player is defeated
-        if (S.adventure.playerHP <= 0) {
-          S.adventure.inCombat = false;
-          S.adventure.combatLog.push('You have been defeated!');
-          log('Defeated in combat! Returning to safety...', 'bad');
-        }
-      }
-    }
-  }
-}
-
-function defeatEnemy() {
-  if (!S.adventure || !S.adventure.currentEnemy) return;
-  
-  // Award experience and loot
-  const enemy = S.adventure.currentEnemy;
-  S.adventure.totalKills++;
-  S.adventure.killsInCurrentArea++;
-  
-  if (!S.adventure.combatLog) S.adventure.combatLog = [];
-  S.adventure.combatLog.push(`${enemy.name} defeated!`);
-  
-  // Check for meat drops
-  if (enemy.drops && enemy.drops.meat && Math.random() < enemy.drops.meat) {
-    const meatGained = 1;
-    S.meat = (S.meat || 0) + meatGained;
-    S.adventure.combatLog.push(`Found ${meatGained} raw meat!`);
-  }
-  
-  // End combat
-  S.adventure.inCombat = false;
-  S.adventure.currentEnemy = null;
-  {
-    const { hp, hpMax } = initHp(0);
-    S.adventure.enemyHP = hp;
-    S.adventure.enemyMaxHP = hpMax;
-  }
-
-  log(`Defeated ${enemy.name}! Kills: ${S.adventure.totalKills}`, 'good');
-
-  // Automatically spawn a new enemy if we're still adventuring and alive
-  if (S.activities.adventure && S.adventure.playerHP > 0) {
-    startAdventureCombat();
-    updateActivityAdventure();
-  }
-}
-
-function startAdventureCombat() {
-  if (!S.adventure || !ADVENTURE_ZONES) return;
-  
-  // Always default to first zone and area for simplicity
-  S.adventure.selectedZone = S.adventure.selectedZone || 0;
-  S.adventure.selectedArea = S.adventure.selectedArea || 0;
-  S.adventure.currentZone = S.adventure.selectedZone;
-  S.adventure.currentArea = S.adventure.selectedArea;
-  
-  const currentZone = ADVENTURE_ZONES[S.adventure.selectedZone];
-  if (!currentZone || !currentZone.areas) return;
-  
-  const currentArea = currentZone.areas[S.adventure.selectedArea];
-  if (!currentArea || !currentArea.enemy) return;
-  
-  // Get enemy type from current area (single enemy per area)
-  const enemyType = currentArea.enemy;
-  const enemyData = ENEMY_DATA[enemyType];
-  
-  if (!enemyData) {
-    log(`Enemy data not found for ${enemyType}`, 'bad');
-    return;
-  }
-  
-  // Start combat
-  S.adventure.inCombat = true;
-  S.adventure.currentEnemy = { ...enemyData, type: enemyType };
-  const { hp, hpMax } = initHp(enemyData.hp);
-  S.adventure.enemyHP = hp;
-  S.adventure.enemyMaxHP = hpMax;
-  S.adventure.playerHP = S.hp; // Sync player HP
-  S.adventure.lastPlayerAttack = 0;
-  S.adventure.lastEnemyAttack = 0;
-  
-  if (!S.adventure.combatLog) S.adventure.combatLog = [];
-  S.adventure.combatLog.push(`A ${enemyData.name} appears!`);
-  
-  log(`Combat started with ${enemyData.name} in ${currentArea.name}!`, 'neutral');
-}
-
-function progressToNextArea() {
-  if (!S.adventure) return;
-  
-  const currentZone = ADVENTURE_ZONES[S.adventure.selectedZone || 0];
-  if (!currentZone || !currentZone.areas) return;
-  
-  const currentArea = currentZone.areas[S.adventure.selectedArea || 0];
-  if (!currentArea) return;
-  
-  // Check if current area is cleared
-  if (S.adventure.killsInCurrentArea < currentArea.killReq) {
-    log('Area not yet cleared! Defeat more enemies first.', 'bad');
-    return;
-  }
-  
-  // Progress to next area
-  if (S.adventure.selectedArea < currentZone.areas.length - 1) {
-    S.adventure.selectedArea++;
-    S.adventure.currentArea = S.adventure.selectedArea;
-    S.adventure.killsInCurrentArea = 0;
-    S.adventure.areasCompleted++;
-    
-    const newArea = currentZone.areas[S.adventure.selectedArea];
-    log(`Advanced to ${newArea.name}!`, 'good');
-  } else if (S.adventure.selectedZone < ADVENTURE_ZONES.length - 1) {
-    // Progress to next zone
-    S.adventure.selectedZone++;
-    S.adventure.currentZone = S.adventure.selectedZone;
-    S.adventure.selectedArea = 0;
-    S.adventure.currentArea = 0;
-    S.adventure.killsInCurrentArea = 0;
-    S.adventure.zonesUnlocked = Math.max(S.adventure.zonesUnlocked, S.adventure.selectedZone + 1);
-    
-    const newZone = ADVENTURE_ZONES[S.adventure.selectedZone];
-    log(`Advanced to ${newZone.name}!`, 'excellent');
-  } else {
-    log('You have reached the end of available content!', 'neutral');
-  }
-  
-  updateActivityAdventure();
-}
-
-function selectArea(areaIndex) {
-  if (!S.adventure) return;
-  
-  S.adventure.selectedArea = areaIndex;
-  S.adventure.currentArea = areaIndex;
-  
-  // Reset kills in current area when switching
-  S.adventure.killsInCurrentArea = 0;
-  
-  // Update the UI
-  updateActivityAdventure();
-  
-  const currentZone = ADVENTURE_ZONES[S.adventure.selectedZone || 0];
-  if (currentZone && currentZone.areas && currentZone.areas[areaIndex]) {
-    const selectedArea = currentZone.areas[areaIndex];
-    log(`Selected area: ${selectedArea.name}`, 'good');
-  }
-}
-
-function selectZone(zoneIndex) {
-  if (!S.adventure) return;
-  
-  S.adventure.selectedZone = zoneIndex;
-  S.adventure.currentZone = zoneIndex;
-  S.adventure.selectedArea = 0; // Reset to first area in new zone
-  S.adventure.currentArea = 0;
-  S.adventure.killsInCurrentArea = 0;
-  
-  // Update the UI
-  updateActivityAdventure();
-  
-  const selectedZone = ADVENTURE_ZONES[zoneIndex];
-  if (selectedZone) {
-    log(`Selected zone: ${selectedZone.name}`, 'good');
-  }
-}
-
-function retreatFromCombat() {
-  if (!S.adventure) return;
-  
-  if (S.adventure.inCombat) {
-    S.adventure.inCombat = false;
-    S.adventure.currentEnemy = null;
-    {
-      const { hp, hpMax } = initHp(0);
-      S.adventure.enemyHP = hp;
-      S.adventure.enemyMaxHP = hpMax;
-    }
-    
-    if (!S.adventure.combatLog) S.adventure.combatLog = [];
-    S.adventure.combatLog.push('You retreated from combat.');
-    
-    log('Retreated from combat safely.', 'neutral');
-  }
-  
-  // Stop adventure activity
-  stopActivity('adventure');
-}
-
-// Missing adventure UI functions - implementing all at once to end bug cycle
-function updateProgressButton() {
-  // Initialize adventure data if needed
-  if (!S.adventure) return;
-  
-  const progressBtn = document.getElementById('progressToNextStageBtn');
-  if (!progressBtn) return;
-  
-  // Check if current area is cleared
-  const currentZone = ADVENTURE_ZONES[S.adventure.selectedZone || 0];
-  if (!currentZone || !currentZone.areas) return;
-  
-  const currentArea = currentZone.areas[S.adventure.selectedArea || 0];
-  if (!currentArea) return;
-  
-  const isAreaCleared = S.adventure.killsInCurrentArea >= currentArea.killReq;
-  
-  progressBtn.disabled = !isAreaCleared;
-  progressBtn.textContent = isAreaCleared ? 'Progress to Next Area' : `Clear Area (${S.adventure.killsInCurrentArea}/${currentArea.killReq})`;
-}
-
-function setupAdventureTabs() {
-  // Setup adventure tab switching if elements exist
-  const tabButtons = document.querySelectorAll('.adventure-tab-btn');
-  
-  tabButtons.forEach(button => {
-    button.onclick = () => {
-      const tabName = button.dataset.tab;
-      
-      // Remove active class from all buttons and contents
-      tabButtons.forEach(btn => btn.classList.remove('active'));
-      document.querySelectorAll('.adventure-tab-content').forEach(content => {
-        content.classList.remove('active');
-        content.style.display = 'none';
-      });
-      
-      // Add active class to clicked button and show content
-      button.classList.add('active');
-      const content = document.getElementById(tabName + 'Tab');
-      if (content) {
-        content.classList.add('active');
-        content.style.display = 'block';
-      }
-      
-      // Update food slots when equipment tab is opened
-      if (tabName === 'equipment') {
-        updateFoodSlots();
-      }
-    };
-  });
-}
 
 
 
@@ -1979,98 +1461,6 @@ function updateMiningRateDisplays() {
   setText('iceRate', `+${iceRate.toFixed(1)}/sec`);
 }
 
-function updateActivityAdventure() {
-  // Initialize adventure data if not exists
-  if (!S.adventure) {
-    const { hp: enemyHP, hpMax: enemyMaxHP } = initHp(0);
-    S.adventure = {
-      currentZone: 0,
-      currentArea: 0,
-      totalKills: 0,
-      areasCompleted: 0,
-      zonesUnlocked: 1,
-      killsInCurrentArea: 0,
-      inCombat: false,
-      playerHP: S.hp,
-      enemyHP,
-      enemyMaxHP,
-      currentEnemy: null,
-      lastPlayerAttack: 0,
-      lastEnemyAttack: 0,
-      combatLog: ['Select an area to begin your adventure...'],
-      selectedZone: 0,
-      selectedArea: 0,
-      bestiary: {} // Track enemy kills for bestiary
-    };
-  }
-
-  // Initialize bestiary if not exists
-  if (!S.adventure.bestiary) {
-    S.adventure.bestiary = {};
-  }
-  
-  // Update location display
-  const currentZone = ADVENTURE_ZONES[S.adventure.selectedZone || S.adventure.currentZone || 0];
-  if (currentZone && currentZone.areas) {
-    const currentArea = currentZone.areas[S.adventure.selectedArea || S.adventure.currentArea || 0];
-    if (currentArea) {
-      setText('currentLocationText', `${currentZone.name} - ${currentArea.name}`);
-      setText('killsRequired', `${S.adventure.killsInCurrentArea}/${currentArea.killReq}`);
-    } else {
-      setText('currentLocationText', 'Unknown Area');
-      setText('killsRequired', '0/0');
-    }
-  } else {
-    setText('currentLocationText', 'Unknown Zone');
-    setText('killsRequired', '0/0');
-  }
-  
-  // Update progress stats
-  setText('totalKills', S.adventure.totalKills);
-  setText('areasCompleted', S.adventure.areasCompleted);
-  setText('zonesUnlocked', S.adventure.zonesUnlocked);
-  
-  // Update combat stats
-  const playerAttack = calculatePlayerCombatAttack();
-  const playerAttackRate = calculatePlayerAttackRate();
-  
-  setText('currentWeapon', 'Fists');
-  const fistBase = 5 + getFistBonuses().damage;
-  setText('baseDamage', fistBase);
-  setText('physiqueDamageBonus', `+${Math.floor((S.stats.physique - 10) * 2)}`);
-  setText('combatAttackRate', playerAttackRate.toFixed(1) + '/s');
-  setText('playerAttack', playerAttack);
-  setText('playerAttackRate', playerAttackRate.toFixed(1) + '/s');
-
-  // Update fist proficiency display
-  updateFistProficiencyDisplay();
-  
-  // Update zone buttons
-  updateZoneButtons();
-  
-  // Update area grid
-  updateAreaGrid();
-  
-  // Update battle display (always visible)
-  updateBattleDisplay();
-  
-  // Update breakthrough progress
-  updateBreakthrough();
-  
-  // Update adventure combat
-  updateAdventureCombat();
-
-  // Update food slots cooldowns
-  updateFoodSlots();
-  
-  // Update sidebar activities
-  updateSidebarActivities();
-  
-  updateProgressButton();
-  
-  // Setup adventure tab switching
-  setupAdventureTabs();
-}
 
 function updateActivityCooking() {
   // Initialize cooking data if not exists
@@ -2271,57 +1661,7 @@ function useFoodSlot(slotNumber) {
   updateAll();
 }
 
-function updateFoodSlots() {
-  if (!S.foodSlots) return;
-  
-  const now = Date.now();
-  const cooldownRemaining = Math.max(0, S.foodSlots.cooldown - (now - S.foodSlots.lastUsed));
-  
-  for (let i = 1; i <= 3; i++) {
-    const slotKey = `slot${i}`;
-    const foodType = S.foodSlots[slotKey];
-    
-    const slotElement = document.getElementById(`foodSlot${i}`);
-    const nameElement = slotElement?.querySelector('.food-name');
-    const iconElement = slotElement?.querySelector('.food-icon');
-    const cooldownElement = document.getElementById(`foodCooldown${i}`);
-    
-    if (!slotElement || !nameElement || !iconElement || !cooldownElement) continue;
-    
-    // Update slot appearance
-    slotElement.classList.remove('empty', 'on-cooldown');
-    
-    if (!foodType) {
-      // Empty slot
-      slotElement.classList.add('empty');
-      iconElement.textContent = '🍽️';
-      nameElement.textContent = 'Empty';
-      cooldownElement.textContent = '';
-    } else {
-      // Food equipped
-      if (foodType === 'meat') {
-        iconElement.textContent = '🥩';
-        nameElement.textContent = `Raw Meat (${S.meat || 0})`;
-      } else if (foodType === 'cookedMeat') {
-        iconElement.textContent = '🍖';
-        nameElement.textContent = `Cooked Meat (${S.cookedMeat || 0})`;
-      }
-      
-      // Show cooldown
-      if (cooldownRemaining > 0) {
-        slotElement.classList.add('on-cooldown');
-        cooldownElement.textContent = `${Math.ceil(cooldownRemaining / 1000)}s`;
-      } else {
-        cooldownElement.textContent = '';
-      }
-    }
-  }
-}
 
-function openFoodManager() {
-  // Switch to cooking activity
-  showActivity('cooking');
-}
 
 
 
@@ -3442,10 +2782,6 @@ function tick(){
   updateAll();
 }
 
-function log(msg, cls=''){
-  const el=document.getElementById('log'); const p=document.createElement('p'); if(cls) p.className=cls; p.textContent=`${timeStr()} ${msg}`; el.appendChild(p); el.scrollTop=el.scrollHeight;
-}
-function timeStr(){ const s=S.time; const h=Math.floor(s/3600), m=Math.floor((s%3600)/60), ss=s%60; return `[${String(h).padStart(2,'0')}:${String(m).padStart(2,'0')}:${String(ss).padStart(2,'0')}]`; }
 
 
 // Activity selector event listeners
@@ -3539,6 +2875,7 @@ function initActivityListeners() {
   // Adventure retreat button event listener
   document.getElementById('retreatButton')?.addEventListener('click', () => {
     retreatFromCombat();
+    stopActivity('adventure');
   });
 }
